+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # DeepDiff Change log
 
+- v8-3-0
+    - Fixed some static typing issues
+    - Added the summarize module for better repr of nested values
+
 - v8-2-0
     - Small optimizations so we don't load functions that are not needed
     - Updated the minimum version of Orderly-set